LayoutTests/platform/blackberry/accessibility/aria-presentational-role-expected.txt

Link and text

This tests that the aria 'presentation' role works by successfully removing the element from the AX tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL firstChild.role should be AXRole: AXLink. Was .
PASS successfullyParsed is true

TEST COMPLETE
